requirements.txt
Django==4.0.6
djangorestframework==3.13.1
django-cors-headers==4.4.0
psycopg2==2.9.3
gunicorn==22.0.0
